+++ test-website/src/Pages/Test/Inputs/SelectOptions.ts
@@ -1,6 +1,4 @@
-﻿import { CustomOptionType } from './CustomOption'
-
-export const colorOptions = [
+﻿export const colorOptions = [
     { value: 'ocean', label: 'Ocean', color: '#00B8D9' },
     { value: 'blue', label: 'Blue', color: '#0052CC', disabled: true },
     { value: 'purple', label: 'Purple', color: '#5243AA' },